+++ src/skeleton/mod.rs
@@ -11,21 +11,21 @@ pub struct Skeleton<'a> {
     pub bones: Vec<Bone<'a>>,
 }
 
-const EXDATA_ID: usize = 0x8000;
+const EXDATA_ID: u32 = 0x8000;
 
 #[derive(Debug, PartialEq)]
 pub struct Bone<'a> {
-    pub id: usize,
+    pub id: u32,
     pub inverse_bind_pose: RowMatrix4<f32>,
     pub name: Cow<'a, str>,
     pub exdata: Option<ExData>,
-    pub parent: Option<usize>,
+    pub parent: Option<u32>,
 }
 
 type ExData = ();
 
 impl Bone<'_> {
-    fn serial_id(&self) -> usize {
+    fn serial_id(&self) -> u32 {
         match self.exdata {
             Some(_) => self.id | EXDATA_ID,
             None => self.id,
